Deprecate std::string interface to std::vector
The source example of usage is using "\xff" literial like binary data
to the disasm function, which takes char* and the size.
However, in C++ it would not corroundspond to a std::string if the
string literial contains any null-byte. It's undefined behavior when a
string containing any null bytes in it. In other words, the compiler or
stdlib cound strip any data after the first occurrance of the null-byte.

Therefore, we deprecate the std::string interface and use
std::vector instead.